+++ Netch/Servers/Shadowsocks/Models/SSD/Main.cs
@@ -12,7 +12,7 @@ namespace Netch.Servers.Shadowsocks.Models.SSD
         /// <summary>
         ///     端口
         /// </summary>
-        public int port;
+        public ushort port;
 
         /// <summary>
         ///     加密方式
